# feat: nested diaries